+++ src/services/oni-save/hooks/useMaterials.ts
@@ -1,24 +1,26 @@
 import * as React from "react";
 import { useSelector, useDispatch } from "react-redux";
+import { SimHashName } from "oni-save-parser";
 
 import { MaterialListItem, materialsSelector } from "../selectors/material";
 import { deleteLooseMaterial } from "../actions/delete-loose-material";
 
 export interface UseMaterials {
   materials: MaterialListItem[];
-  onDeleteLooseMaterial(): void;
 }
 
-export default function useMaterials(): UseMaterials {
-  const dispatch = useDispatch();
+export function useMaterialList(): MaterialListItem[] {
   const materials = useSelector(materialsSelector);
+  return materials;
+}
 
-  const onDeleteLooseMaterial = React.useCallback(() => {
-    dispatch(deleteLooseMaterial());
-  }, [dispatch]);
-
-  return {
-    materials,
-    onDeleteLooseMaterial
-  }
+export function useDeleteMaterials() {
+  const dispatch = useDispatch();
+  const deleteMaterial = React.useCallback(
+    (materialType?: SimHashName) => {
+      dispatch(deleteLooseMaterial(materialType));
+    },
+    [dispatch]
+  );
+  return deleteMaterial;
 }