+++ src/services/oni-save/materials.ts
@@ -1,0 +1,5 @@
+import { SimHashNames } from "oni-save-parser";
+
+// TODO: Seeds, clothing, other sweepables
+export const MaterialGameObjectNames = [...SimHashNames];
+export type MaterialObjectName = ArrayValues<typeof MaterialGameObjectNames>;