+++ src/types/types.d.ts
@@ -1,11 +1,9 @@
-type Omit<T, K> = Pick<T, Exclude<keyof T, K>>;
-
 type DeepPartial<T> = {
   [P in keyof T]?: T[P] extends Array<infer U>
     ? Array<DeepPartial<U>>
     : T[P] extends ReadonlyArray<infer U>
     ? ReadonlyArray<DeepPartial<U>>
-    : DeepPartial<T[P]>
+    : DeepPartial<T[P]>;
 };
 
 type ArrayValues<T> = T extends Array<infer U> ? U : never;